requirements.txt
streamlit
langchain
chromadb
python-dotenv